+++ lib/platform-bible-react/src/components/advanced/inventory/inventory-utils.ts
